fix(THEEDGE-3739): move columns and fetch to edge utilities
